add leaderboard + create game but doesn't update database

+++ site-project/users/models.py
@@ -5,7 +5,7 @@ from PIL import Image
 class Profile(models.Model):
 	user = models.OneToOneField(User, on_delete=models.CASCADE)
 	image = models.ImageField(default='default.jpg', upload_to='profile_pics')
-	elo = models.IntegerField(default='1000')
+	elo = models.FloatField(default='1000')
 	victory = models.IntegerField(default='0')
 	draw = models.IntegerField(default='0')
 	loose = models.IntegerField(default='0')
